# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: Dart Backend API
  version: 0.5.0
  description: Server-authoritative API for Dart | for you. Identity sessions use Secure HttpOnly cookies and CSRF tokens.
servers:
  - url: /api/v1
...
paths:
...
  /admin/auth/email/start:
    post:
      tags: [Staff Auth]
      summary: Request a six-digit verification code for an allowed dashboard email
      description: Always returns a generic response when email delivery is configured so callers cannot enumerate the Staff allowlist.
      operationId: startStaffEmailAccess
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffEmailStart" }
      responses:
        "202": { description: Generic verification challenge response }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /admin/auth/email/resend:
    post:
      tags: [Staff Auth]
      summary: Request a replacement verification code for an allowed dashboard email
      operationId: resendStaffEmailAccess
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffEmailStart" }
      responses:
        "202": { description: Generic verification challenge response }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /admin/auth/email/verify:
    post:
      tags: [Staff Auth]
      summary: Verify the one-time email code and issue the secure Dart Staff session
      operationId: verifyStaffEmailAccess
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffEmailVerify" }
      responses:
        "200": { $ref: "#/components/responses/Authenticated" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /admin/staff:
    get:
      tags: [Staff Auth]
      summary: List Staff accounts, pending email access entries and permission catalogue
      operationId: listStaffAccess
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Staff access directory }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/staff/allowlist:
    post:
      tags: [Staff Auth]
      summary: Allow a dashboard email and assign Owner or Staff role
      operationId: allowStaffEmail
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffEmailAllowance" }
      responses:
        "201": { description: Dashboard email added to the PostgreSQL allowlist }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /admin/staff/allowlist/{id}:
    delete:
      tags: [Staff Auth]
      summary: Remove an unclaimed Staff email access entry
      operationId: removeStaffEmailAccess
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Pending allowlist entry removed }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/staff/{id}/access:
    patch:
      tags: [Staff Auth]
      summary: Activate or disable Staff access and revoke old sessions
      operationId: setStaffAccessStatus
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Staff access state updated }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/staff/{id}/sessions/revoke:
    post:
      tags: [Staff Auth]
      summary: Revoke every Dart session for one Staff account
      operationId: revokeStaffSessions
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Staff sessions revoked }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dart_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
...
    StaffEmailStart:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
    StaffEmailVerify:
      type: object
      required: [challengeId, code]
      properties:
        challengeId: { type: string, format: uuid }
        code: { type: string, pattern: "^[0-9]{6}$", writeOnly: true }
    StaffEmailAllowance:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        role: { type: string, enum: [owner, staff] }
        permissionKeys:
          type: array
          maxItems: 300
          items: { type: string }
...
    Profile:
      type: object
      required: [id, accountType, code, name, email, phones, status]
      properties:
        id: { type: string, format: uuid }
        accountType: { type: string, enum: [customer, staff, representative] }
        code: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        phones: { type: array, items: { type: string } }
        birthday: { type: [string, "null"], format: date }
        status: { type: string }
        emailVerified: { type: boolean }
        mustChangePassword: { type: boolean }
    Authenticated:
      type: object
      required: [user, csrfToken]
      properties:
        user: { $ref: "#/components/schemas/Profile" }
        csrfToken: { type: string, writeOnly: true }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
    Authenticated:
      description: Authenticated session created or replaced
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Authenticated" }
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: Duplicate or stale state conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Request or business validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServiceUnavailable:
      description: Required secure integration is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }